requirements.txt
astroid==2.0.4
autopep8==1.4.3
Click==7.0
colorama==0.4.0
Flask==1.0.2
Flask-MySQLdb==0.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
memory-profiler==0.54.0
mysqlclient==1.3.13
neurtu==0.2.0
numpy==1.15.4
pandas==0.23.4
psutil==5.4.8
pycodestyle==2.4.0
pylint==2.1.1
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
tqdm==4.28.1
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
